requirements.txt
certifi==2018.11.29
chardet==3.0.4
click==8.0.4
colorama==0.4.5
configparser==5.2.0
dataclasses==0.8
Flask==2.0.3
Flask-Login==0.5.0
flask-paginate==2024.4.12
gunicorn==21.2.0
ibm-db==3.0.1
idna==2.8
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.3
pymqi==1.9.2
pyparsing==3.0.7
python-dotenv==0.20.0
requests==2.21.0
typing-extensions==4.1.1
urllib3==1.24.1
Werkzeug==2.0.3
zipp==3.6.0
